+++ src/Asset/EncoreExtension.php
@@ -1,0 +1,35 @@
+<?php
+
+/*
+ * Copyright (c) 2023 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\WatchlistBundle\Asset;
+
+use HeimrichHannot\EncoreContracts\EncoreEntry;
+use HeimrichHannot\EncoreContracts\EncoreExtensionInterface;
+use HeimrichHannot\WatchlistBundle\HeimrichHannotWatchlistBundle;
+
+class EncoreExtension implements EncoreExtensionInterface
+{
+    /**
+     * {@inheritDoc}
+     */
+    public function getBundle(): string
+    {
+        return HeimrichHannotWatchlistBundle::class;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getEntries(): array
+    {
+        return [
+            EncoreEntry::create('contao-watchlist-bundle', 'src/Resources/assets/js/contao-watchlist-bundle-init.js')
+                ->addJsEntryToRemoveFromGlobals('contao-watchlist-bundle'),
+        ];
+    }
+}